refactor wrappers, add render with rgb_array
